+++ assets/js/src/public/analytics/segment.js
@@ -1,18 +1,19 @@
 /**
- * @module Analytics Tracking
- * @description Allow Facebook Pixel and Google Analytics tracking on specified events.
+ * @module Segment Analytics Tracking
+ * @description Allow Facebook Pixel and Google Analytics tracking on specified Segment events.
  */
 
 import delegate from 'delegate';
 import { STORE_DOMAIN } from 'publicConfig/wp-settings';
-import * as tools from '../../utils/tools';
+import * as tools from 'utils/tools';
+import { on } from 'utils/events';
 
 const el = {
 	segment: tools.getNodes('bc-segment-tracker')[0],
 };
 
-const handleAddToCartTracker = () => {
-	const cartTrigger = tools.getNodes('[data-tracking-event="add_to_cart"]', false, document, true)[0];
+const handleAddToCartTracker = (e) => {
+	const cartTrigger = e ? e.detail.cartButton : tools.getNodes('[data-tracking-event="add_to_cart_message"]', false, document, true)[0];
 
 	if (!cartTrigger) {
 		return;
@@ -22,7 +23,7 @@ const handleAddToCartTracker = () => {
 	const jsonData = JSON.parse(analyticsData);
 
 	analytics.track('Product Added', {
-		cart_id: jsonData.cart_id,
+		cart_id: e ? e.detail.cart_id : jsonData.cart_id,
 		product_id: jsonData.product_id,
 		variant: jsonData.variant_id,
 	});
@@ -61,6 +62,8 @@ const bindEvents = () => {
 	tools.getNodes('bc-product-quick-view-content', true, document).forEach((dialog) => {
 		delegate(dialog, '.bc-product__title-link', 'click', handleClickTracker);
 	});
+
+	on(document, 'bigcommerce/analytics_trigger', handleAddToCartTracker);
 };
 
 const init = () => {