+++ assets/js/src/public/buttons/index.js
@@ -5,12 +5,10 @@
 
 import quickViewDialog from './quick-view-dialog';
 import pagination from './pagination';
-import analytics from './analytics';
 
 const init = () => {
 	quickViewDialog();
 	pagination();
-	analytics();
 };
 
 export default init;