+++ src/BigCommerce/Assets/Theme/JS_Config.php
@@ -3,6 +3,9 @@
 
 namespace BigCommerce\Assets\Theme;
 
+use BigCommerce\Taxonomies\Channel\Channel;
+use BigCommerce\Taxonomies\Channel\Connections;
+
 class JS_Config {
 	/**
 	 * @var array
@@ -14,8 +17,14 @@ class JS_Config {
 	 */
 	private $directory;
 
-	public function __construct( $asset_directory ) {
-		$this->directory       = trailingslashit( $asset_directory );
+	/**
+	 * @var Connections
+	 */
+	private $connections;
+
+	public function __construct( $asset_directory, $connections ) {
+		$this->directory   = trailingslashit( $asset_directory );
+		$this->connections = $connections;
 	}
 
 	public function get_data() {
@@ -28,10 +37,24 @@ class JS_Config {
 						'not_available' => __( 'The selected product combination is currently unavailable.', 'bigcommerce' ),
 					],
 				],
+				'channel'      => $this->get_current_channel_data(),
 			];
 			$this->data = apply_filters( 'bigcommerce/js_config', $this->data );
 		}
 
 		return $this->data;
 	}
+
+	private function get_current_channel_data() {
+		$current_channel = $this->connections->current();
+		if ( $current_channel ) {
+			return [
+				'id'   => get_term_meta( $current_channel->term_id, Channel::CHANNEL_ID, true ),
+				'name' => $current_channel->name,
+			];
+		}
+
+		return false;
+	}
+
 }