+++ src/BigCommerce/Container/Assets.php
@@ -7,6 +7,7 @@ namespace BigCommerce\Container;
 use BigCommerce\Assets\Admin;
 use BigCommerce\Assets\Theme;
 use BigCommerce\Plugin;
+use BigCommerce\Taxonomies\Channel\Connections;
 use Pimple\Container;
 
 class Assets extends Provider {
@@ -81,7 +82,7 @@ class Assets extends Provider {
 		};
 
 		$container[ self::FRONTEND_CONFIG ] = function ( Container $container ) {
-			return new Theme\JS_Config( $container[ self::PATH ] );
+			return new Theme\JS_Config( $container[ self::PATH ], new Connections() );
 		};
 
 		$container[ self::FRONTEND_LOCALIZATION ] = function( Container $container ) {