+++ src/BigCommerce/Schema/Reviews_Table.php
@@ -4,38 +4,42 @@
 namespace BigCommerce\Schema;
 
 
+/**
+ * Class Reviews_Table
+ *
+ * @deprecated since 4.0. Reviews are fetched dynamically with a cache stored in post meta
+ */
 class Reviews_Table extends Table_Maker {
 	const NAME = 'bc_reviews';
 
-	protected $schema_version = '3.0-dev2';
+	protected $schema_version = '4.0';
 
 	protected $tables = [ self::NAME ];
 
-	protected function get_table_definition( $table ) {
-
-		global $wpdb;
-		$table_name       = $wpdb->$table;
-		$charset_collate  = $wpdb->get_charset_collate();
-		switch ( $table ) {
-			case self::NAME:
-				return "CREATE TABLE {$table_name} (
-				        review_id BIGINT(20) unsigned NOT NULL,
-				        post_id BIGINT(20) unsigned DEFAULT 0,
-				        bc_id BIGINT(20) unsigned NOT NULL,
-				        title VARCHAR(255) NOT NULL DEFAULT '',
-				        content LONGTEXT NOT NULL DEFAULT '',
-				        status ENUM( 'approved', 'disapproved', 'pending' ) NOT NULL,
-				        rating TINYINT(1) unsigned NOT NULL DEFAULT 0,
-				        author_email VARCHAR(255) NOT NULL DEFAULT '',
-				        author_name VARCHAR(255) NOT NULL DEFAULT '',
-				        date_reviewed DATETIME NOT NULL,
-				        date_created DATETIME NOT NULL,
-				        date_modified DATETIME NOT NULL,
-				        PRIMARY KEY  (review_id),
-				        KEY bc_id (bc_id),
-				        KEY status (status, date_reviewed, bc_id),
-				        KEY rating (rating)
-				        ) $charset_collate";
+	/**
+	 * Override for the table registration,
+	 * as it was removed in version 4.0
+	 *
+	 * @return void
+	 */
+	public function register_tables() {
+		// drop the tables
+		if ( $this->schema_update_required() ) {
+			foreach ( $this->tables as $table ) {
+				$this->drop_table( $table );
+			}
+			$this->mark_schema_update_complete();
 		}
+	}
+
+	private function drop_table( $table ) {
+		/** @var \wpdb $wpdb */
+		global $wpdb;
+		$blog_table = $wpdb->prefix . $table;
+		$wpdb->query( "DROP TABLE IF EXISTS `$blog_table`" );
+	}
+
+	protected function get_table_definition( $table ) {
+		return '';
 	}
 }